Clean up class_parsing and class_processing
